Clean up unused test navigations and rename packages
