+++ src/ucsd_f1tenth_gazebo/launch/main.launch
@@ -26,7 +26,7 @@
   <!-- Manual control -->
   <!-- <include file="$(find ucsd_f1tenth_navigation)/launch/teleop.launch" /> -->
 
-  <!-- <include file="$(find ucsd_f1tenth_navigation_v2)/launch/follow_waypoints.launch" /> -->
+  <include file="$(find ucsd_f1tenth_navigation)/launch/waypoint_follower.launch" />
 
 
 </launch>
