+++ src/ucsd_f1tenth_navigation/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.0.2)
-project(ucsd_f1tenth_navigation_v2)
+project(ucsd_f1tenth_navigation)
 
 find_package(catkin REQUIRED COMPONENTS
   rospy
@@ -10,6 +10,9 @@ catkin_package()
 include_directories(${catkin_INCLUDE_DIRS})
 
 catkin_install_python(PROGRAMS
-  scripts/follow_waypoints.py
+  scripts/create_global_path.py
+  scripts/joy_teleop.py
+  scripts/waypoint_follower.py
+  scripts/waypoint_logger.py
   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
