--- src/ucsd_f1tenth_navigation/config/costmap_common_params.yaml
@@ -1,30 +1,0 @@
-
-#---standard pioneer footprint---
-#---(in meters)---
-#footprint: [ [0.254, -0.0508], [0.1778, -0.0508], [0.1778, -0.1778], [-0.1905, -0.1778], [-0.254, 0], [-0.1905, 0.1778], [0.1778, 0.1778], [0.1778, 0.0508], [0.254, 0.0508] ]
-footprint: [ [-0.05,-0.0625], [0.25,-0.0625], [0.25,0.0625], [-0.05,0.0625] ]
-
-
-transform_tolerance: 0.2
-map_type: costmap
-
-obstacle_layer:
- enabled: true
- obstacle_range: 3.0
- raytrace_range: 3.5
- inflation_radius: 0.2
- track_unknown_space: false
- combination_method: 1
-
- observation_sources: laser_scan_sensor
- laser_scan_sensor: {data_type: LaserScan, topic: scan, marking: true, clearing: true}
-
-
-inflation_layer:
-  enabled:              true
-  cost_scaling_factor:  10.0  # exponential rate at which the obstacle cost drops off (default: 10)
-  inflation_radius:     0.5  # max. distance from an obstacle at which costs are incurred for planning paths.
-
-static_layer:
-  enabled:              true
-  map_topic:            "/map"
