--- src/ucsd_f1tenth_navigation/config/global_costmap_params.yaml
@@ -1,16 +1,0 @@
-global_costmap:
-  global_frame: map
-  robot_base_frame: base_link
-  update_frequency: 1.0
-  publish_frequency: 0.5
-  static_map: true
- 
-  transform_tolerance: 0.5
-  plugins:
-    - {name: static_layer,            type: "costmap_2d::StaticLayer"}
-    - {name: obstacle_layer,          type: "costmap_2d::VoxelLayer"}
-    - {name: inflation_layer,         type: "costmap_2d::InflationLayer"}
-
-
-
-
