--- src/ucsd_f1tenth_navigation/config/global_path_svg.txt
@@ -1,1 +1,0 @@
-M2427.5 46.9998C2103.5 0.49984 432.5 -35.5 165 69.5C-102.5 174.5 -9.49985 854.5 208.5 987.5C426.5 1120.5 1007.24 535.804 1447.5 582C1861.73 625.465 2103 1042 2427.5 1008.5C2752 975 2870 716.5 2870 518C2870 319.5 2751.5 93.4999 2427.5 46.9998Z
