--- src/ucsd_f1tenth_navigation/config/local_costmap_params.yaml
@@ -1,15 +1,0 @@
-local_costmap:
-  global_frame: map
-  robot_base_frame: base_link
-  update_frequency: 5.0
-  publish_frequency: 2.0
-  static_map: false
-  rolling_window: true
-  width: 5.5
-  height: 5.5
-  resolution: 0.1
-  transform_tolerance: 0.5
-  
-  plugins:
-   - {name: static_layer,        type: "costmap_2d::StaticLayer"}
-   - {name: obstacle_layer,      type: "costmap_2d::ObstacleLayer"}
