--- src/ucsd_f1tenth_navigation/launch/follow_waypoints.launch
@@ -1,49 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-  <arg name="namespace" default="/navigation"/>
-  <arg name="cmd_timeout" default="0.5"/>
-  <arg name="starting_x" default="0"/>
-  <arg name="starting_y" default="0"/>
-
-  <!-- Vehicle pose -->
-  <arg name="ackermann_cmd_topic" default="/vehicle/ackermann_cmd" />
-  <arg name="odom_topic" default="/vehicle/odom" />
-  <arg name="cmd_vel_topic" default="/vehicle/cmd_vel" />
-  <arg name="new_wp_topic" default="/waypoints/goal" />
-
-  <arg name="global_path_csv" default="$(find ucsd_f1tenth_navigation)/config/ucsd_west_waypoints.csv"/>
-
-
-  <group ns="$(arg namespace)">
-
-	  <!--  **************** Visualisation ****************  -->
-	  <node name="rviz" pkg="rviz" type="rviz" args="-d $(find ucsd_f1tenth_navigation_v2)/rviz/navigation.rviz"/>
-    <!-- Nav sat -->
-    <include file="$(find ucsd_f1tenth_navigation)/launch/start_navsat.launch" />
-
-    <!-- Run the ekf for map to odom config -->
-    <node pkg="robot_localization"
-          type="ekf_localization_node"
-          name="ekf_localization_with_gps">
-      <rosparam command="load"
-                file="$(find ucsd_f1tenth_navigation)/config/robot_localization.yaml" />
-    </node>
-
-    <!-- Run the map server -->
-    <include file="$(find ucsd_f1tenth_navigation)/launch/start_map_server.launch" />
-    <!-- Convert cmd_vel to ackermann drive -->
-    <node name="follow_waypoints" pkg="ucsd_f1tenth_navigation_v2"
-          type="follow_waypoints.py" >
-      <param name="ackermann_cmd_topic" value="$(arg ackermann_cmd_topic)" />
-      <param name="odom_topic" value="$(arg odom_topic)" />
-      <param name="cmd_vel_topic" value="$(arg cmd_vel_topic)" />
-      <param name="new_wp_topic" value="$(arg new_wp_topic)" />
-      <param name="starting_x" value="$(arg starting_x)" />
-      <param name="starting_y" value="$(arg starting_y)" />
-    </node>
-    <node pkg="ucsd_f1tenth_navigation" type="create_global_path.py" name="create_global_path" >
-      <param name="global_path_csv" value="$(arg global_path_csv)"/>
-      <!-- <param name="global_path_svg" value="$(arg global_path_svg)"/> -->
-    </node>
-  </group>
-</launch>
