+++ src/ucsd_f1tenth_navigation/launch/navsat_transform.launch
@@ -1,0 +1,23 @@
+<?xml version="1.0"?>
+<launch>
+  <node pkg="robot_localization" type="navsat_transform_node" name="navsat_transform_node" respawn="true">
+
+    <param name="magnetic_declination_radians" value="0"/>
+    <param name="yaw_offset" value="0"/>
+    <param name="zero_altitude" value="true"/>
+
+    <param name="broadcast_cartesian_transform" value="false"/>
+    <param name="publish_filtered_gps" value="false"/>
+    <param name="predict_to_current_time" value="true" />
+
+
+    <param name="use_odometry_yaw" value="false"/>
+    <param name="wait_for_datum" value="false"/>
+
+    <remap from="/imu/data" to="/gazebo/imu/data" />
+    <remap from="/gps/fix" to="/gazebo/gps/fix" />
+    <remap from="/odometry/filtered" to="/vehicle/odom" />
+
+  </node>
+
+</launch>
