--- src/ucsd_f1tenth_navigation/launch/start_map_server.launch
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-    <arg name="map_file" default="$(find ucsd_f1tenth_navigation)/maps/ucsd_west_map.yaml"/>
-    <!-- <arg name="map_file" default="$(find ucsd_f1tenth_navigation)/maps/empty_map.yaml"/> -->
-    <node name="map_server" pkg="map_server" type="map_server" args="$(arg map_file)" />
-</launch>
