+++ src/ucsd_f1tenth_navigation/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>ucsd_f1tenth_navigation_v2</name>
+  <name>ucsd_f1tenth_navigation</name>
   <version>0.0.0</version>
-  <description>The ucsd_f1tenth_navigation_v2 package</description>
+  <description>The ucsd_f1tenth_navigation package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
