--- src/ucsd_f1tenth_navigation/paths/ucsd_track.csv
@@ -1,2 +1,0 @@
-
-
