+++ src/ucsd_f1tenth_navigation/scripts/create_global_path.py
@@ -1,11 +1,8 @@
 #!/usr/bin/env python
-import math
-import numpy as np
 import pandas as pd
 import rospy
 from nav_msgs.msg import Path
-from geometry_msgs.msg import Pose, PoseStamped
-from tf.transformations import quaternion_from_euler
+from geometry_msgs.msg import PoseStamped
 
 
 class GlobalPathCreator:
@@ -81,76 +78,9 @@ class GlobalPathCreator:
         ----------------
         Returns: None
         """
-
         for i in range(len(waypoints)):
             pose = self.make_pose(waypoints.iloc[i])
             self.global_path.poses.append( pose )
-
-
-class SVGPathCreator:
-    def __init__(self):
-        self.global_path_topic = "/global_path"
-        global_path_svg = rospy.get_param("~global_path_svg", "global_path_svg.txt")
-
-        # create path publisher
-        self.path_pub = rospy.Publisher(self.global_path_topic, Path, queue_size=1)
-
-        parsed_points = self.parse_svg(global_path_svg)
-        global_path = self.create_path(parsed_points)
-
-        while not rospy.is_shutdown():
-            self.path_pub.publish(global_path)
-
-    def parse_svg(self, svg_path: str):
-        with open(svg_path) as svg:
-            path = svg.readline()
-            path = path.replace("M", " ")
-            path = path.replace("C", " ")
-            path = path.replace("Z", " ")
-
-            path_list = path.split(" ")
-            path_list = path_list[1: -1]
-
-            path_array = np.array(path_list, dtype=float)
-            path_array = path_array.reshape((int(path_array.shape[0]/2), 2))
-            path_array = np.flip(path_array)
-        return path_array
-
-    def create_path(self, points):
-        global_path = Path()
-        global_path.header.frame_id = "/map"
-        global_path.header.stamp = rospy.Time.now()
-
-        for i, point in enumerate(points):
-            pose = PoseStamped()
-            pose.header.frame_id = "/map"
-            pose.header.stamp = rospy.Time.now()
-
-            # 0.25 inches per pixel * 0.0254 meters per inch
-            pixel_to_meters = 0.25 * 0.0254
-            pose.pose.position.x = point[0] * pixel_to_meters
-            pose.pose.position.y = point[1] * pixel_to_meters
-            pose.pose.position.z = 0
-
-            if i + 1 == len(points):
-                q = self.get_orientation(point, points[0])
-            else:
-                q = self.get_orientation(point, points[i+1])
-
-            # set orientation of pose
-            pose.pose.orientation.x = q[0]
-            pose.pose.orientation.y = q[1]
-            pose.pose.orientation.z = q[2]
-            pose.pose.orientation.w = q[3]
-
-            global_path.poses.append(pose)
-
-        return global_path
-
-    def get_orientation(self, point_1, point_2):
-        angle = math.atan2(point_2[1] - point_1[1], point_2[0] - point_1[0])
-        q = quaternion_from_euler(0, 0, angle)
-        return q
 
 
 
@@ -158,7 +88,6 @@ if __name__ == "__main__":
     try:
         rospy.init_node("global_path_creator", anonymous=True)
         GlobalPathCreator()
-        # SVGPathCreator()
         rospy.spin()
     except rospy.ROSInterruptException:
         pass
